refactor: move link component to common
