+++ src/app/snippets/page.tsx
@@ -2,8 +2,7 @@ import type { Snippet } from 'contentlayer/generated'
 import { allSnippets } from 'contentlayer/generated'
 import type { Metadata } from 'next'
 
-import { Container, EmptyState } from '@/components/common'
-import Link from '@/components/link'
+import { Container, EmptyState, Link } from '@/components/common'
 import PageHeader from '@/components/page-header'
 import { ROUTES } from '@/constants/links'
 import cn from '@/lib/cn'
