+++ src/components/mdx/image.tsx
@@ -7,7 +7,7 @@ import { useState } from 'react'
 import cn from '@/lib/cn'
 import { getDomainFromUrl } from '@/lib/utils'
 
-import Link from '../link'
+import { Link } from '../common'
 import ImageZoom from './image-zoom'
 
 type BaseImageProps = Omit<NextImageProps, 'width' | 'height' | 'fill'>
