+++ src/components/navigation.tsx
@@ -10,9 +10,8 @@ import cn from '@/lib/cn'
 import { defaultMetadata } from '@/lib/metadata'
 
 import Logo from '../../assets/images/logo.svg'
-import { Button, Container } from './common'
+import { Button, Container, Link } from './common'
 import { Hamburger } from './icons'
-import Link from './link'
 import ThemeSwitch from './theme-switch'
 
 const animation = {
